API - check app config before dropping database
